Move findItem and selectStyle out
